% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-rect-flow.R
\name{geom_rect_flow}
\alias{geom_rect_flow}
\title{geom_rect_flow}
\usage{
geom_rect_flow(mapping = NULL, asp = 1.5, width = 0.8, ...)
}
\arguments{
\item{mapping}{aesthetic mapping}

\item{asp}{aspect ration of rectangle box (width vs hight)}

\item{width}{width of the rectangles}

\item{...}{additional parameters passed to 'geom_rect'}
}
\description{
draw rectangle box for flow chart
}
\author{
Guangchuang Yu
}
