\name{ggflow-package}
\alias{ggflow-package}
\alias{ggflow}
\docType{package}
\title{
2D scatter plots of a flowFrame object with ggplot2 
}
\description{
This package reads flowFrame object and plots them in beautiful scatter plots using ggplot2. You can also select/gate cells/events on a plot to remove/select relevant cells and you can display a visible, rectangular gate with percentage of selected cells highlighted. Makes plotting flow cytometry data easy and pretty. 
}
\details{
\tabular{ll}{
Package: \tab ggflow\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-10-16\cr
License: GPLv3
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
Francesco Vallania
fvallani@stanford.edu
}

~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
