% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggflow.R
\name{ggflow}
\alias{ggflow}
\title{ggflow}
\usage{
ggflow(
  data,
  layout = igraph::layout_as_tree,
  start_shave = 0.3,
  end_shave = 0.3,
  type_shave = "proportion",
  ...
)
}
\arguments{
\item{data}{a data frame containing two columns indicating 'from->to' relationship}

\item{layout}{flow chart layout}

\item{start_shave}{offset of the line segment (the from end)}

\item{end_shave}{offset of the line segment (the to end)}

\item{type_shave}{type of the 'start_shave' and 'end_shave', one of 'proportion' or 'distance'}

\item{...}{additional parameters passed to 'ggtree'}
}
\description{
draw flow chart
}
\examples{
library(ggplot2)
library(ggflow)

d <- data.frame(
  from = c("A", "A", "A", "B", "C", "F"),
  to = c("B", "C", "D", "E", "F", "G")
)
ggflow(d, arrow=arrow(type='closed')) + 
    geom_rect_flow(fill=NA, color='black') +
    geom_text(aes(label=label), size=20)
}
\author{
Guangchuang Yu
}
