% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.r
\docType{import}
\name{reexports}
\alias{reexports}
\alias{geom_scatter_rect}
\alias{arrow}
\alias{ggplot}
\alias{aes}
\alias{geom_text}
\alias{geom_segment}
\alias{ggsave}
\alias{\%<+\%}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{ggfun}{\code{\link[ggfun:attacher]{\%<+\%}}, \code{\link[ggfun]{geom_scatter_rect}}}

  \item{ggplot2}{\code{\link[ggplot2]{aes}}, \code{\link[ggplot2]{geom_segment}}, \code{\link[ggplot2]{geom_text}}, \code{\link[ggplot2]{ggplot}}, \code{\link[ggplot2]{ggsave}}}

  \item{grid}{\code{\link[grid]{arrow}}}
}}

